- reorganized assets - implement barebone ecs system - optimized light engine, logger, mesh engine, texture, etc

+++ Minecraft Clone/src/Level/Entity/Entity.cpp
@@ -10,6 +10,7 @@ Entity::Entity() = default;
 Entity::~Entity() = default;
 
 void Entity::Tick(Dimension* dimension) {
-  dimension->context_.entities_list_->entity_type_list_[properties_.type_]
+  dimension->context_.entities_list_
+      ->entity_type_list_[static_cast<int>(properties_.type_)]
       ->Tick(this, dimension);
 }
